feat: add base websocket handling
