+++ package.json
@@ -2,11 +2,11 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "index.ts",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "node --loader ts-node/esm index.ts",
+        "start:dev": "nodemon --exec node --loader ts-node/esm index.ts",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
@@ -16,6 +16,8 @@
         "nodemon": "^2.0.16"
     },
     "dependencies": {
-        "ws": "^8.8.0"
+        "@types/ws": "^8.5.12",
+        "ts-node": "^10.9.2",
+        "ws": "^8.18.0"
     }
 }
